+++ Sources/ChessKitEngine/EngineType.swift
@@ -9,11 +9,13 @@ import ChessKitEngineCore
 public enum EngineType: Int {
     
     case stockfish
+    case lc0
     
     /// Internal mapping from Swift to Obj-C type.
     var objc: EngineType_objc {
         switch self {
         case .stockfish:    return .stockfish
+        case .lc0:          return .lc0
         }
     }
     
@@ -21,6 +23,7 @@ public enum EngineType: Int {
     public var name: String {
         switch self {
         case .stockfish:    return "Stockfish"
+        case .lc0:          return "LeelaChessZero (Lc0)"
         }
     }
     
@@ -28,6 +31,34 @@ public enum EngineType: Int {
     public var version: String {
         switch self {
         case .stockfish:    return "15.1"
+        case .lc0:          return "0.29"
+        }
+    }
+    
+    /// Engine-specific options to configure at initialization.
+    var setupCommands: [EngineCommand] {
+        switch self {
+        case .stockfish:
+            let evalFile = Bundle.module.url(forResource: "nn-1337b1adec5b", withExtension: "nnue")?
+                .absoluteString
+                .replacingOccurrences(of: "file://", with: "")
+            
+            return [
+                evalFile != nil ?
+                    .setoption(id: "EvalFile", value: evalFile!) : nil,
+                .setoption(id: "Use NNUE", value: "true"),
+                .setoption(id: "UCI_AnalyseMode", value: "true")
+            ].compactMap { $0 }
+        case .lc0:
+            let weights = Bundle.module.url(forResource: "192x15_network", withExtension: nil)?
+                .absoluteString
+                .replacingOccurrences(of: "file://", with: "")
+            
+            return [
+                .setoption(id: "Backend", value: "eigen"),
+                weights != nil ?
+                    .setoption(id: "WeightsFile", value: weights!) : nil
+            ].compactMap { $0 }
         }
     }